backend/requirements.txt
# Core dependencies
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
aiofiles==23.2.1

# API and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
websockets==12.0
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
openpyxl==3.1.2
xlsxwriter==3.1.9

# AI/ML Core
scikit-learn==1.3.2
scipy==1.11.4

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Deep Learning
torch==2.2.0
transformers==4.36.2
sentence-transformers==2.2.2

# XGBoost & Gradient Boosting
xgboost==2.0.3
lightgbm==4.2.0

# NLP
konlpy==0.6.0
jpype1==1.4.1

# Optimization
cvxpy==1.4.1
pulp==2.7.0

# MLOps
mlflow==2.9.2
optuna==3.5.0
shap==0.44.0

# Imbalanced data
imbalanced-learn==0.11.0

# Scheduling and workflow
croniter==2.0.1
apscheduler==3.10.4

# GraphQL (for OwnerClan integration)
gql==3.4.1
requests-toolbelt==1.0.0

# Authentication
pyjwt==2.8.0
cryptography==41.0.7

# Monitoring
psutil==5.9.6
gputil==1.4.0

# Version control
gitpython==3.1.40
packaging==23.2

# Email (for alerts)
aiosmtplib==3.0.1

# Logging and monitoring
colorlog==6.8.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Plotting (optional for backend reports)
matplotlib==3.7.3
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1
pre-commit==3.6.0